MISTRAL-main/requirements.txt
chromadb>=0.4.22
langchain>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
sentence-transformers>=2.3.0
pypdf>=3.17.0
python-docx>=1.1.0
requests>=2.31.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
